+++ ecrire/bootstrap/config/initial.php
@@ -1,8 +1,12 @@
 <?php
 
+use function SpipLeague\Component\Kernel\param;
+
 /**
  * Constantes de démarrage de SPIP
  */
+
+$ecrire = param('spip.dirs.core');
 
 /**
  * Indique que SPIP est chargé
@@ -13,14 +17,11 @@
  */
 define('_ECRIRE_INC_VERSION', '1');
 
-if (!defined('_DIR_RESTREINT_ABS')) {
-	/** le nom du repertoire ecrire/ */
-	define('_DIR_RESTREINT_ABS', 'ecrire/');
-}
-
-/** Chemin relatif pour aller dans ecrire
- *  vide si on est dans ecrire, 'ecrire/' sinon */
-define('_DIR_RESTREINT', (is_dir(_DIR_RESTREINT_ABS) ? _DIR_RESTREINT_ABS : ''));
+/**
+ * Chemin relatif pour aller dans ecrire
+ * vide si on est dans ecrire, 'ecrire/' sinon
+ */
+define('_DIR_RESTREINT', (is_dir($ecrire) ? $ecrire : ''));
 
 /** Chemin relatif pour aller à la racine */
 define('_DIR_RACINE', _DIR_RESTREINT ? '' : '../');