+++ spip.php
@@ -7,14 +7,18 @@
  * Arnaud Martin, Antoine Pitrou, Philippe Rivière, Emmanuel Saint-James
  *
  * Ce programme est un logiciel libre distribué sous licence GNU/GPL.
- * Pour plus de détails voir le fichier COPYING.txt ou l'aide en ligne.
+ * Pour plus de détails voir le fichier LICENSE ou l'aide en ligne.
  */
 
-# ou est l'espace prive ?
-if (!defined('_DIR_RESTREINT_ABS')) {
-	define('_DIR_RESTREINT_ABS', 'ecrire/');
-}
-include_once _DIR_RESTREINT_ABS.'inc_version.php';
+use function SpipLeague\Component\Kernel\param;
+
+require_once __DIR__ . '/vendor/autoload.php';
+
+$ecrire = param('spip.dirs.core');
+
+include_once $ecrire . 'inc_version.php';
+/** @todo Replace by $app = boot() call  */
 
 # au travail...
-include _DIR_RESTREINT_ABS.'public.php';
+include $ecrire . 'public.php';
+/** @todo Replace by $app->run() */